Refactor ticket management logic into its own files
