+++ ext/semian/resource.h
@@ -6,9 +6,9 @@ Functions here are associated with rubyland operations.
 #ifndef SEMIAN_RESOURCE_H
 #define SEMIAN_RESOURCE_H
 
-#include "semian.h" // FIXME remove this once configure_tickets has been refactored
 #include "types.h"
 #include "sysv_semaphores.h"
+#include "tickets.h"
 
 // Ruby variables
 ID id_timeout;
