+++ ext/semian/semian.h
@@ -19,15 +19,8 @@ Implements Init_semian, which is used as C/Ruby entrypoint.
 #include <ruby/io.h>
 
 //semian includes
-#include "types.h"
 #include "resource.h"
-#include "sysv_semaphores.h" // FIXME: TEMPORARY
 
 void Init_semian();
-
-// FIXME: These declarations will be modified in a subsequent PR
-// temporarily placed here while refactoring
-void
-configure_tickets(int sem_id, int tickets, int should_initialize);
 
 #endif //SEMIAN_H
